Update to record completion time in time frame format
