+++ plugins/woocommerce-admin/client/tasks/fills/use-record-completion-time.ts
@@ -4,13 +4,18 @@
 import { useRef } from '@wordpress/element';
 import { recordEvent } from '@woocommerce/tracks';
 
+/**
+ * Internal dependencies
+ */
+import { getTimeFrame } from '~/utils';
+
 const useRecordCompletionTime = ( taskName: string, startTime?: number ) => {
 	const _startTime = useRef( startTime || window.performance.now() );
 
 	const recordCompletionTime = () => {
 		recordEvent( 'task_completion_time', {
 			task_name: taskName,
-			time: window.performance.now() - _startTime.current,
+			time: getTimeFrame( window.performance.now() - _startTime.current ),
 		} );
 	};
