+++ plugins/woocommerce-admin/client/utils/index.js
@@ -59,3 +59,32 @@ export const sift = ( arr, partitioner ) =>
 		},
 		[ [], [] ]
 	);
+
+const timeFrames = [
+	{ name: '0-2s', max: 2 },
+	{ name: '2-5s', max: 5 },
+	{ name: '5-10s', max: 10 },
+	{ name: '10-15s', max: 15 },
+	{ name: '15-20s', max: 20 },
+	{ name: '20-30s', max: 30 },
+	{ name: '30-60s', max: 60 },
+	{ name: '>60s' },
+];
+
+/**
+ * Returns time frame for a given time in milliseconds.
+ *
+ * @param {number} timeInMs - time in milliseconds
+ *
+ * @return {string} - Time frame.
+ */
+export const getTimeFrame = ( timeInMs ) => {
+	for ( const timeFrame of timeFrames ) {
+		if ( ! timeFrame.max ) {
+			return timeFrame.name;
+		}
+		if ( timeInMs < timeFrame.max * 1000 ) {
+			return timeFrame.name;
+		}
+	}
+};
